backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.6.0
pymongo==4.9.2
pydantic[email]==2.7.1
python-dotenv==1.0.1
